Keep demo root actions in root controllers
