+++ demos/social-auth/app/ui/home/page.tsx
@@ -1,10 +1,10 @@
 import type { ExternalProviderLink } from '../../utils/external-auth.ts'
-import { AuthCard } from '../../ui/auth-card.tsx'
-import { Document } from '../../ui/document.tsx'
-import { TextField } from '../../ui/form-field.tsx'
-import { EmailIcon, PasswordIcon } from '../../ui/icons.tsx'
-import { Notice } from '../../ui/notice.tsx'
-import * as styles from '../../ui/styles.ts'
+import { AuthCard } from '../auth-card.tsx'
+import { Document } from '../document.tsx'
+import { TextField } from '../form-field.tsx'
+import { EmailIcon, PasswordIcon } from '../icons.tsx'
+import { Notice } from '../notice.tsx'
+import * as styles from '../styles.ts'
 import { ExternalAuthSection } from './external-auth-section.tsx'
 import { LoginFooter } from './footer.tsx'
