+++ demos/unpkg/app/ui/package-browser/file-content.ts
