+++ demos/unpkg/app/ui/package-browser/format-bytes.ts
